Test viewDate in separate file
I think the tests are leaking and it's causing tests to fail, in the
current test file the tests has to be in a certain order - otherwise
they will fail.

One solution is to put them in separate files (like in this commit), but
of course the optimal solution would be to fix the leakage.
