+++ test/tests.spec.js
@@ -1175,53 +1175,4 @@ describe('Datetime', () => {
 		});
 
 	});
-
-	describe('with viewDate', () => {
-    it('date value', () => {
-      const date = new Date(2000, 0, 15, 2, 2, 2, 2),
-          strDate = moment(date).format('MMMM YYYY'),
-          component = utils.createDatetime({ viewDate: date });
-      expect(utils.getViewDateValue(component)).toEqual(strDate);
-    });
-
-    it('moment value', () => {
-      const date = new Date(2000, 0, 15, 2, 2, 2, 2),
-          mDate = moment(date),
-          strDate = mDate.format('MMMM YYYY'),
-          component = utils.createDatetime({ viewDate: mDate });
-      expect(utils.getViewDateValue(component)).toEqual(strDate);
-    });
-
-    it('string value', () => {
-      const date = new Date(2000, 0, 15, 2, 2, 2, 2),
-          mDate = moment(date),
-          strDate = mDate.format('L') + ' ' + mDate.format('LT'),
-          expectedStrDate = mDate.format('MMMM YYYY'),
-          component = utils.createDatetime({ viewDate: strDate });
-      expect(utils.getViewDateValue(component)).toEqual(expectedStrDate);
-    });
-
-    it('UTC value from UTC string', () => {
-      const date = new Date(2000, 0, 15, 2, 2, 2, 2),
-          momentDateUTC = moment.utc(date),
-          strDateUTC = momentDateUTC.format('L') + ' ' + momentDateUTC.format('LT'),
-          expectedStrDate = momentDateUTC.format('MMMM YYYY'),
-          component = utils.createDatetime({ viewDate: strDateUTC, utc: true });
-      expect(utils.getViewDateValue(component)).toEqual(expectedStrDate);
-    });
-
-    it('invalid string value', () => {
-      const strDate = 'invalid string',
-          expectedStrDate = moment().format('MMMM YYYY'),
-          component = utils.createDatetime({ viewDate: strDate });
-      expect(utils.getViewDateValue(component)).toEqual(expectedStrDate);
-    });
-
-    it('invalid moment object', () => {
-      const mDate = moment(null),
-          expectedStrDate = moment().format('MMMM YYYY'),
-          component = utils.createDatetime({ viewDate: mDate });
-      expect(utils.getViewDateValue(component)).toEqual(expectedStrDate);
-    });
-	});
 });
